Seperate mapping_utils file for mapping helper methods
